requirements.txt
pandas==1.5.0
requests==2.28.1
boto3==1.24.89
psycopg2-binary==2.9.4